+++ tests/unit_tests/testcases/app/test_qmarkdowntextedit.h
@@ -10,6 +10,7 @@ class TestQMarkdownTextEdit : public QObject {
     void testUpFromStartOfSecondLineMovesToFirstLine();
     void testToPlainTextPreservesNoBreakSpaces();
     void testRCodeBlockHighlighting();
+    void testLinkedCheckBoxDetectionInReadOnlyEditor();
 };
 
 #endif    // TESTQMARKDOWNTEXTEDIT_H